refactor: 타입 폴더 생성 및 스케줄 인터페이스 생성 #102
