+++ front/app/_types/user/Dog.ts
